refactor: User Section in own file
